Move `hash-graph-client(-generator)` to `libs/@local`
