+++ apps/hash-api/package.json
@@ -22,8 +22,8 @@
     "@blockprotocol/core": "0.0.12",
     "@blockprotocol/type-system": "0.0.3-canary-20230111084611",
     "@graphql-tools/schema": "8.5.0",
-    "@hashintel/hash-graph-client": "0.0.0-private",
     "@local/hash-backend-utils": "0.0.0-private",
+    "@local/hash-graph-client": "0.0.0-private",
     "@local/hash-isomorphic-utils": "0.0.0-private",
     "@local/hash-subgraph": "0.0.0-private",
     "@opentelemetry/api": "1.3",
