+++ libs/@local/hash-graph-client/.eslintrc.cjs
