+++ libs/@local/hash-graph-client/.gitignore
