+++ libs/@local/hash-graph-client/.npmignore
