+++ libs/@local/hash-graph-client/.openapi-generator/VERSION
