+++ libs/@local/hash-graph-client/LICENSE.md
