+++ libs/@local/hash-graph-client/base.ts
