+++ libs/@local/hash-graph-client/common.ts
