+++ libs/@local/hash-graph-client/configuration.ts
