+++ libs/@local/hash-graph-client/tsconfig.json
