+++ libs/@local/hash-isomorphic-utils/package.json
@@ -24,7 +24,7 @@
   "dependencies": {
     "@apollo/client": "3.6.9",
     "@blockprotocol/core": "0.0.12",
-    "@hashintel/hash-graph-client": "0.0.0-private",
+    "@local/hash-graph-client": "0.0.0-private",
     "@local/hash-subgraph": "0.0.0-private",
     "@sentry/browser": "7.30.0",
     "immer": "9.0.6",
