+++ tests/hash-backend-integration/jest.config.cjs
@@ -13,9 +13,9 @@ module.exports = {
       "<rootDir>/../../libs/@local/hash-backend-utils/src$1",
     "@local/hash-isomorphic-utils(.*)":
       "<rootDir>/../../libs/@local/hash-isomorphic-utils/src$1",
-    "@local/hash-subgraph(.*)": "<rootDir>/../../libs/@local/hash-subgraph$1",
-    "@hashintel/hash-graph-client":
+    "@local/hash-graph-client":
       "<rootDir>/../../apps/hash-graph/clients/typescript",
+    "@local/hash-subgraph(.*)": "<rootDir>/../../libs/@local/hash-subgraph$1",
   },
   setupFiles: ["@local/hash-backend-utils/environment"],
   testMatch: [
