+++ tests/hash-backend-integration/package.json
@@ -14,8 +14,8 @@
   "dependencies": {
     "@apps/hash-api": "0.0.0-private",
     "@blockprotocol/type-system": "0.0.3-canary-20230111084611",
-    "@hashintel/hash-graph-client": "0.0.0-private",
     "@local/hash-backend-utils": "0.0.0-private",
+    "@local/hash-graph-client": "0.0.0-private",
     "@local/hash-isomorphic-utils": "0.0.0-private",
     "@local/hash-subgraph": "0.0.0-private",
     "execa": "5.1.1",
